+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 - BREAKING: `Terminal:send` takes options as a table instead of ordered parameters.
+- `Terminal:send` accepts `decorator` option as a string to look up by name.
 
 ## [0.0.2] - 2025-07-19